requirements.txt
flask==3.1.0
requests==2.32.3
schedule==1.2.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
dnspython
